+++ src/venus/vkr_device_memory.h
@@ -36,6 +36,6 @@ vkr_device_memory_release(struct vkr_device_memory *mem);
 int
 vkr_device_memory_export_fd(struct vkr_device_memory *mem,
                             VkExternalMemoryHandleTypeFlagBits handle_type,
-                            int *out_fd);
+                            uintptr_t *out_fd);
 
 #endif /* VKR_DEVICE_MEMORY_H */
